// packages/components/nodes/memory/AgentMemory/MySQLAgentMemory/mysqlSaver.ts
import { BaseCheckpointSaver, Checkpoint, CheckpointMetadata } from '@langchain/langgraph'
import { RunnableConfig } from '@langchain/core/runnables'
import { BaseMessage } from '@langchain/core/messages'
import { DataSource } from 'typeorm'
import { CheckpointTuple, SaverOptions, SerializerProtocol } from '../interface'
import { IMessage, MemoryMethods } from '../../../../src/Interface'
import { mapChatMessageToBaseMessage } from '../../../../src/utils'

export class MySQLSaver extends BaseCheckpointSaver implements MemoryMethods {
    protected isSetup: boolean
    config: SaverOptions
    threadId: string
    tableName = 'checkpoints'

     {
        
        this.config = config
        const { threadId } = config
        this.threadId = threadId
    }

    : string {
        // Trim and normalize case, turn whitespace into underscores
        tableName = tableName.t.t.

        // Val
        ) {
            th
        }

        return tableName
    }

    p: Promise<DataSource> {
        const { datasourceOptions } = this.config
         {
            th
        }
        // Prevent using default Postgres port, otherwise will throw uncaught error and crashing the app
         {
            th
        }
        
        awa
        return dataSource
    }

    p: Promise<void> {
         return

        try {
            
            
            await queryRunner.manager.query(`
                CREATE TABLE IF NOT EXISTS ${tableName} (
                    th NOT NULL,
                     NOT NULL,
                    pa,
                    checkpoint BLOB,
                    metadata BLOB,
                    PRIMARY KEY (th
                );`)
            awa
        }  {
            
            th
        }

        this.isSetup = true
    }

    a: Promise<CheckpointTuple | undefined> {
        
        awa

        const thread_id = config.configurable?.thread_id || this.threadId
        const checkpoint_id = config.configurable?.checkpoint_id
        

        try {
            
            const sql = checkpoint_id
                ? `SELECT checkpoint, parent_id, metadata FROM ${tableName} WHERE thread_id = ? AND checkpoint_id = ?`
                : `SELECT thread_id, checkpoint_id, parent_id, checkpoint, metadata FROM ${tableName} WHERE thread_id = ? ORDER BY checkpoint_id DESC LIMIT 1`

            
            awa

             {
                const row = rows[0]
                return {
                    config: {
                        configurable: {
                            thread_id: row.thread_id || thread_id,
                            checkpoint_id: row.checkpoint_id || checkpoint_id
                        }
                    },
                    )) as Checkpoint,
                    meta)) as CheckpointMetadata,
                    parentConfig: row.parent_id
                        ? {
                              configurable: {
                                  thread_id,
                                  checkpoint_id: row.parent_id
                              }
                          }
                        : undefined
                }
            }
        }  {
            
            th
        } finally {
            awa
        }
        return undefined
    }

    a: AsyncGenerator<CheckpointTuple, void, unknown> {
        
        awa
        
        try {
            const threadId = config.configurable?.thread_id || this.threadId
            
            let sql = `SELECT thread_id, checkpoint_id, parent_id, checkpoint, metadata FROM ${tableName} WHERE thread_id = ? ${
                before ? 'AND checkpoint_id < ?' : ''
            } ORDER BY checkpoint_id DESC`
             {
                sql += ` LIMIT ${limit}`
            }
            

            
            awa

             {
                f {
                    yield {
                        config: {
                            configurable: {
                                thread_id: row.thread_id,
                                checkpoint_id: row.checkpoint_id
                            }
                        },
                        )) as Checkpoint,
                        meta)) as CheckpointMetadata,
                        parentConfig: row.parent_id
                            ? {
                                  configurable: {
                                      thread_id: row.thread_id,
                                      checkpoint_id: row.parent_id
                                  }
                              }
                            : undefined
                    }
                }
            }
        }  {
            
            th
        } finally {
            awa
        }
    }

    a: Promise<RunnableConfig> {
        
        awa

         return {}
        try {
            
            const row = [
                config.configurable?.thread_id || this.threadId,
                checkpoint.id,
                config.configurable?.checkpoint_id,
                Buffe), // Encode to binary
                Buffe) // Encode to binary
            ]
            

            
                           VALUES (?, ?, ?, ?, ?)
                           ON , meta`

            awa
            awa
        }  {
            
            th
        } finally {
            awa
        }

        return {
            configurable: {
                thread_id: config.configurable?.thread_id || this.threadId,
                checkpoint_id: checkpoint.id
            }
        }
    }

    a: Promise<void> {
         return

        
        awa
        

        try {
            
            const query = `DELETE FROM ${tableName} WHERE thread_id = ?;`
            awa
            awa
        }  {
            
        } finally {
            awa
        }
    }

    async getChatMessages(
        overrideSessionId = '',
        returnBaseMessages = false,
        prependMessages?: IMessage[]
    ): Promise<IMessage[] | BaseMessage[]> {
         return []

        .find({
            where: {
                sessionId: overrideSessionId,
                chatflowid: this.config.chatflowid
            },
            order: {
                createdDate: 'ASC'
            }
        })

         {
            
        }

         {
            
        }

        let returnIMessages: IMessage[] = []
        f {
            returnIMessages.push({
                message: m.content as string,
                type: m.role
            })
        }

        return returnIMessages
    }

    a: Promise<void> {
        // Empty as it's not being used
    }

    a: Promise<void> {
         return
        awa
    }
}
